libc: arm: a15: add optimized memcpy implementation
This optimization is extracted from cortex-strings and bionic-ized, then
applied to Cortex-A15 cpu. This optimization requires NEON for armv7-a or
vfpv2 for armv6. NEON is optional in cortex-a9, so I did not change
cortex-a9.mk.

This new implementation does not have __memcpy_chk(), so __memcpy_chk.cpp
is included for arch which uses the new implementation.

Change-Id: I57cdaf8dba7f80ff35d8a4292e9de99beee53ef2